Move containers to relevant layout
